Refactor for CodeClimate
