+++ extension/src/lsp-client/language-client.ts
@@ -1,32 +1,16 @@
-import path from 'path'
 import { ExtensionContext, workspace } from 'vscode'
 import {
   LanguageClient,
-  LanguageClientOptions,
-  ServerOptions,
-  TransportKind
+  LanguageClientOptions
 } from 'vscode-languageclient/node'
-import { Disposable } from './class/dispose'
+import { ServerModule } from './server-options'
+import { Disposable } from '../class/dispose'
 
 export class DVCLanguageClient extends Disposable {
   private client: LanguageClient
 
   constructor(context: ExtensionContext) {
     super()
-
-    const serverModule = context.asAbsolutePath(
-      path.join('..', 'languageServer', 'dist', 'server.js')
-    )
-    const debugOptions = { execArgv: ['--nolazy', '--inspect=6009'] }
-
-    const serverOptions: ServerOptions = {
-      run: { module: serverModule, transport: TransportKind.ipc },
-      debug: {
-        module: serverModule,
-        transport: TransportKind.ipc,
-        options: debugOptions
-      }
-    }
 
     const clientOptions: LanguageClientOptions = {
       documentSelector: [
@@ -54,12 +38,17 @@ export class DVCLanguageClient extends Disposable {
     this.client = new LanguageClient(
       'dvc-vscode-lsp',
       'DVC Language Server',
-      serverOptions,
+      new ServerModule(context).options,
       clientOptions
     )
 
     // Start the client. This will also launch the server
+    this.dispose.track(this.start())
+  }
+
+  start() {
     this.client.start()
+    return this
   }
 
   stop() {
